requirements.txt
aiofiles==23.2.1
aiogram==3.11.0
aiohttp==3.9.5
aiosignal==1.4.0
annotated-types==0.7.0
asyncio==4.0.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
click==8.3.0
colorama==0.4.6
Flask==3.1.2
flask_cors==6.0.1
frozenlist==1.7.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
magic-filter==1.0.12
MarkupSafe==3.0.3
multidict==6.6.4
mysql-connector-python==9.4.0
pillow==11.3.0
propcache==0.3.2
pydantic==2.8.2
pydantic_core==2.20.1
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
typing_extensions==4.15.0
Werkzeug==3.1.3
yarl==1.20.1
sqlalchemy==2.0.44


